Move 'insights.web' properties to 'spring.http'
Relocate the 'spring.insights.web.log-request-details' property to
'spring.http.log-request-details'.

Closes gh-14313
