--- spring-boot-project/spring-boot-autoconfigure/src/main/java/org/springframework/boot/autoconfigure/insight/InsightsProperties.java
@@ -1,54 +1,0 @@
-/*
- * Copyright 2012-2018 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.boot.autoconfigure.insight;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-/**
- * Configuration properties for customizing application insights.
- *
- * @author Brian Clozel
- * @since 2.1.0
- */
-@ConfigurationProperties(prefix = "spring.insights")
-public class InsightsProperties {
-
-	private final Web web = new Web();
-
-	public Web getWeb() {
-		return this.web;
-	}
-
-	public static class Web {
-
-		/**
-		 * Whether logging of (potentially sensitive) request details at DEBUG and TRACE
-		 * level is allowed.
-		 */
-		private boolean logRequestDetails;
-
-		public boolean isLogRequestDetails() {
-			return this.logRequestDetails;
-		}
-
-		public void setLogRequestDetails(boolean logRequestDetails) {
-			this.logRequestDetails = logRequestDetails;
-		}
-
-	}
-
-}
