+++ spring-boot-project/spring-boot-docs/src/main/asciidoc/using-spring-boot.adoc
@@ -788,7 +788,7 @@ Spring WebFlux applications, developer tools will enable DEBUG logging for the
 Spring Framework web infrastructure. This will give you information about the incoming
 request, which handler is processing it, the response outcome, etc. If you wish to log
 all request details (including potentially sensitive information), you can turn on
-the `spring.insights.web.log-request-details` configuration property.
+the `spring.http.log-request-details` configuration property.
 
 TIP: For a complete list of the properties that are applied by the devtools, see
 {sc-spring-boot-devtools}/env/DevToolsPropertyDefaultsPostProcessor.{sc-ext}[DevToolsPropertyDefaultsPostProcessor].
